Split out eth1 interface and ethers implementation
